server/requirements.txt
# requirements.txt

fastapi==0.115.6
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.2.0
azure-core==1.32.0
loguru==0.7.3
uvicorn==0.32.1
python-dotenv==1.0.1
websockets==10.0
openai==1.59.3
jsonpath_ng